watchdog: improve the daily archiver to include partial updates and tunable num images per day. Also add a buffer for the last N images.

+++ watchdog.py
@@ -46,6 +46,11 @@ class Archiver:
         self.day_buffer_dir = os.path.join(self.options.data_dir, 'tmp_day_buffer')
         if not os.path.isdir(self.day_buffer_dir):
             os.makedirs(self.day_buffer_dir)
+
+        self.recent_buffer_dir = os.path.join(self.options.data_dir, 'tmp_recent_buffer')
+        if not os.path.isdir(self.recent_buffer_dir):
+            os.makedirs(self.recent_buffer_dir)
+
         self.previousTime = None
         self.fakePreviousNow = None
 
@@ -62,16 +67,24 @@ class Archiver:
                 self.previousTime = now.replace(hour=hour, minute=minute, second=second)
                 print ("[Debug] previous time was ", self.previousTime)
 
-    def flushDay(self, tmpDayDir, year, month, day):
+    def flushDay(self, tmpDayDir, year, month, day, isPartial):
         imageFiles = sorted(glob.glob(tmpDayDir + '/*.jpg'))
         if len(imageFiles) > 0:
             outputDir = os.path.join(self.options.data_dir, 'days')
             if not os.path.isdir(outputDir):
                 os.makedirs(outputDir)
-            mp4Filename = "{:04d}_{:02d}_{:02d}.mp4".format(year, month, day)
+            mp4Filename = None
+            if isPartial:
+                mp4Filename = "{:04d}-{:02d}-{:02d}_partial_{}.mp4".format(year, month, day, len(imageFiles))
+            else:
+                mp4Filename = "{:04d}-{:02d}-{:02d}.mp4".format(year, month, day)
+                partialFiles = glob.glob(outputDir + '/{:04d}-{:02d}-{:02d}_partial*.mp4'.format(year, month, day))
+                for f in partialFiles:
+                    os.remove (f)
             outputMp4 = os.path.join(outputDir, mp4Filename)
             createMp4(imageFiles, outputMp4)
-        shutil.rmtree(tmpDayDir)
+        if (not isPartial):
+            shutil.rmtree(tmpDayDir)
 
     def maybeFlushPreviousDays(self, now):
         dayFolders = os.listdir(self.day_buffer_dir)
@@ -80,7 +93,7 @@ class Archiver:
             fullPath = os.path.join(self.day_buffer_dir, dirName)
             if not os.path.isdir(fullPath):
                 continue
-            m = re.match("(\d\d\d\d)_(\d\d)_(\d\d)", dirName)
+            m = re.match("(\d\d\d\d)-(\d\d)-(\d\d)", dirName)
             if not m:
                 continue
             year = int(m.group(1))
@@ -88,16 +101,10 @@ class Archiver:
             day = int(m.group(3))
             if year == now.year and month == now.month and day == now.day:
                 continue 
-            self.flushDay(fullPath, year, month, day)
+            self.flushDay(fullPath, year, month, day, isPartial=False)
 
-    def processImage(self, image):
-        now = datetime.now()
-        # Option: simulate a super fast clock to see the behavior over days.
-        # if self.fakePreviousNow:
-        #     now = self.fakePreviousNow + timedelta(minutes=240+30)
-        # self.fakePreviousNow = now
-
-        imageDirName = now.strftime("%Y_%m_%d")
+    def handleDayBuffer(self, now, image):
+        imageDirName = now.strftime("%Y-%m-%d")
         imageDir = os.path.join(self.day_buffer_dir, imageDirName)
 
         # try to guess the previousTime from a previous run,
@@ -107,7 +114,9 @@ class Archiver:
         if self.previousTime and not isSameDay(now, self.previousTime):
             self.maybeFlushPreviousDays(now)
 
-        if self.previousTime and isSameHour(now, self.previousTime):
+        minDelta = timedelta(seconds=((24*3600.)/self.options.num_images_per_day))
+
+        if self.previousTime and (now - self.previousTime) < minDelta:
             return
             
         self.previousTime = now
@@ -118,7 +127,28 @@ class Archiver:
         imageName = now.strftime("%H_%M_%S.jpg")
         imPath = os.path.join(imageDir, imageName)
         cv.imwrite(imPath, image)
-        print ("[Debug] Wrote ", imPath)
+        self.flushDay(imageDir, now.year, now.month, now.day, isPartial=True)
+
+    def handleRecentBuffer(self, now, image):
+        imageName = now.strftime("%Y-%m-%d_%H_%M_%S.jpg")
+        imPath = os.path.join(self.recent_buffer_dir, imageName)
+        cv.imwrite(imPath, image)
+        images = sorted (os.listdir(self.recent_buffer_dir))
+        if (len(images) > self.options.recent_buffer_size):
+            toRemove = images[0:len(images)-self.options.recent_buffer_size]
+            for f in toRemove:
+                os.remove(os.path.join(self.recent_buffer_dir, f))
+
+    def processImage(self, image):
+        now = datetime.now()
+
+        # Option: simulate a super fast clock to see the behavior over days.
+        # if self.fakePreviousNow:
+        #     now = self.fakePreviousNow + timedelta(minutes=5)
+        # self.fakePreviousNow = now
+
+        self.handleDayBuffer (now, image)        
+        self.handleRecentBuffer (now, image)
 
 class WatchDog:
     def __init__(self, options):
@@ -164,6 +194,8 @@ def parseCommandLine():
     parser.add_argument('server_url', help='Server address and port in zmq format. Example: "tcp://myserver.com:4242"')
     parser.add_argument('--image-server-password', help='Password to connect to the image server')
     parser.add_argument('--data-dir', help='Directory used to save images and alerts', default='data')
+    parser.add_argument('--recent-buffer-size', help='Number of images to keep in the recent buffer', type=int, default=60)
+    parser.add_argument('--num-images-per-day', help='Number of images in the daily summary (default is 4 per hour)', type=int, default=24*4)
     return parser.parse_args()
 
 if __name__ == "__main__":
